core/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
idna==3.8
Markdown==3.7
modeltranslation==0.25
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
